Fix to un-symlink files with no .bak
If the original did not exist we would still symlink the file, but the
reset task was not de-symlinking it because there was not .bak created.
Now we remove the symlink regardless of whether there was an original.

This commit also inlines safe_symlink2 as it only made the tasks more
complex to understand.